# src/AppForSEII2526.Web/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Compra/GetCompra": {
      "get": {
        "tags": [
          "Compra"
        ],
        "operationId": "GetCompra",
        "parameters": [
          {
            "name": "Id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compra/CreateCompra": {
      "post": {
        "tags": [
          "Compra"
        ],
        "operationId": "CreateCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraDetailDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvío",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "apellidoCliente": {
            "minLength": 1,
            "type": "string"
          },
          "nombreCliente": {
            "minLength": 1,
            "type": "string"
          },
          "precioTotal": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "number",
            "format": "float"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          },
          "direccionEnvío": {
            "minLength": 1,
            "type": "string"
          },
          "metodoDePago": {
            "$ref": "#/components/schemas/metodoDePago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          },
          "correoElectonico": {
            "type": "string",
            "nullable": true
          },
          "teléfono": {
            "type": "integer",
            "format": "int32"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CompraForCreateDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvío",
          "nombreCliente"
        ],
        "type": "object",
        "properties": {
          "apellidoCliente": {
            "minLength": 1,
            "type": "string"
          },
          "nombreCliente": {
            "minLength": 1,
            "type": "string"
          },
          "precioTotal": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "number",
            "format": "float"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          },
          "direccionEnvío": {
            "minLength": 1,
            "type": "string"
          },
          "metodoDePago": {
            "$ref": "#/components/schemas/metodoDePago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          },
          "correoElectonico": {
            "type": "string",
            "nullable": true
          },
          "teléfono": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CompraItemDTO": {
        "type": "object",
        "properties": {
          "material": {
            "type": "string",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "precio": {
            "maximum": 2147483647,
            "minimum": 0,
            "type": "number",
            "format": "float"
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "metodoDePago": {
        "enum": [
          "TarjetaCredito",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      },
...
    }
  }
}